+++ CHANGES.md
@@ -13,6 +13,7 @@
 - Removed `archiveOutput` key, modify the target file location through explicit move
 - Removed `publishing` configuration
 - Added `testSources` and `testResources` to simplify setting of these properties
+- Generated files are now stored in a separate directory in ./build/generated
 
 # 0.13 2020-06-23
 - Add utility functions for source file and classpath generation, see the `wemi.generation` package